Add files via upload
Write a program that will compute the weight of an object on the other planets or on the moon.  Your program will allow the user to enter the weight on Earth.  Your program will allow the user to enter the planet from a list of options.
In order to calculate the weight on another planet you need to know the gravitational pull of the other planets according to this chart:

Mercury = 0.38 
Venus = 0.91
Earth's Moon = 0.17 
Mars = 0.38
Jupiter = 2.36 
Saturn = 0.91 
Uranus = 0.89
Neptune = 1.13
Pluto = 0.07

To calculate the weight on another planet multiply the weight on the Earth by the gravitational pull of the planet/moon as in the chart.  For example, if you weigh 100 pounds on Earth you will weigh 17 pounds on the moon (100 * 0.17)